+++ apps/backend/services/users/api/client.go
@@ -12,5 +12,5 @@ import (
 //go:generate go tool with-modfile mockery --name Client --structname API --outpkg testing --output ./testing --filename generated__users_api_mocks.go
 type Client interface {
 	RegisterUser(ctx context.Context, newUser usersservice.NewUser) (*users.User, error)
-	GetUser(ctx context.Context, ID int64) (*users.User, error)
+	GetUser(ctx context.Context, ID int32) (*users.User, error)
 }
