+++ apps/backend/services/users/internal/users/get_user.go
@@ -9,8 +9,8 @@ import (
 	"github.com/mdobak/go-xerrors"
 )
 
-func (s *Service) GetUser(ctx context.Context, ID int64) (*users.User, error) {
-	fetchedUsers, err := s.storage.ListUsers(ctx, storage.ListUsersParam{ID: []int64{ID}})
+func (s *Service) GetUser(ctx context.Context, ID int32) (*users.User, error) {
+	fetchedUsers, err := s.storage.ListUsers(ctx, storage.ListUsersParam{ID: []int32{ID}})
 	if err != nil {
 		return nil, xerrors.Newf("failed to list users: %w", err)
 	}
